rewrite of NetworkAPI (fuze URLApi and NetworkAPI)
