+++ src/main/java/me/srrapero720/watermedia/api/network/models/twitter/GuestTokenResponse.java
@@ -1,0 +1,5 @@
+package me.srrapero720.watermedia.api.network.models.twitter;
+
+public class GuestTokenResponse {
+    public String guest_token;
+}
