+++ src/main/java/me/srrapero720/watermedia/api/network/models/twitter/RequestDetails.java
@@ -1,0 +1,8 @@
+package me.srrapero720.watermedia.api.network.models.twitter;
+
+import java.util.Map;
+
+public class RequestDetails {
+    public Map<String, Boolean> variables;
+    public Map<String, Boolean> features;
+}
