--- src/main/java/me/srrapero720/watermedia/api/network/twitter/GuestTokenResponse.java
@@ -1,5 +1,0 @@
-package me.srrapero720.watermedia.api.network.twitter;
-
-class GuestTokenResponse {
-    public String guest_token;
-}
