--- src/main/java/me/srrapero720/watermedia/api/network/twitter/RequestDetails.java
@@ -1,8 +1,0 @@
-package me.srrapero720.watermedia.api.network.twitter;
-
-import java.util.Map;
-
-class RequestDetails {
-    public Map<String, Boolean> variables;
-    public Map<String, Boolean> features;
-}
