+++ src/main/java/me/srrapero720/watermedia/api/player/impl/VLCMusicPlayer.java
@@ -1,0 +1,4 @@
+package me.srrapero720.watermedia.api.player.impl;
+
+public class VLCMusicPlayer {
+}
