+++ src/main/java/me/srrapero720/watermedia/api/player/impl/VLCPlayer.java
@@ -1,0 +1,4 @@
+package me.srrapero720.watermedia.api.player.impl;
+
+public abstract class VLCPlayer {
+}
