+++ src/main/java/me/srrapero720/watermedia/api/player/impl/VLCVideoPlayer.java
@@ -1,0 +1,4 @@
+package me.srrapero720.watermedia.api.player.impl;
+
+public class VLCVideoPlayer {
+}
