--- src/main/java/me/srrapero720/watermedia/api/url/fixers/TikTokFixer.java
@@ -1,5 +1,0 @@
-package me.srrapero720.watermedia.api.url.fixers;
-
-public class TikTokFixer {
-
-}
