--- src/main/java/me/srrapero720/watermedia/api/url/fixers/special/SpecialFixer.java
@@ -1,5 +1,0 @@
-package me.srrapero720.watermedia.api.url.fixers.special;
-
-import me.srrapero720.watermedia.api.url.fixers.URLFixer;
-
-public abstract class SpecialFixer extends URLFixer {}
