+++ src/main/resources/META-INF/services/me.srrapero720.watermedia.api.WaterMediaAPI
@@ -1,5 +1,7 @@
 me.srrapero720.watermedia.api.player.PlayerAPI
 me.srrapero720.watermedia.api.cache.CacheAPI
 me.srrapero720.watermedia.api.image.ImageAPI
-me.srrapero720.watermedia.api.url.UrlAPI
+me.srrapero720.watermedia.api.config.ConfigAPI
+me.srrapero720.watermedia.api.image.ImageAPI
+me.srrapero720.watermedia.api.network.NetworkAPI
 me.srrapero720.watermedia.api.rendering.RenderAPI
