--- src/main/resources/META-INF/services/me.srrapero720.watermedia.api.url.fixers.URLFixer
@@ -1,10 +1,0 @@
-# NORMAL FIXERS
-me.srrapero720.watermedia.api.url.fixers.OnedriveFixer
-me.srrapero720.watermedia.api.url.fixers.DriveFixer
-me.srrapero720.watermedia.api.url.fixers.YoutubeFixer
-me.srrapero720.watermedia.api.url.fixers.TwitchFixer
-me.srrapero720.watermedia.api.url.fixers.TwitterFixer
-me.srrapero720.watermedia.api.url.fixers.DropboxFixer
-
-# SPECIAL FIXERS
-me.srrapero720.watermedia.api.url.fixers.special.PHFixer
